Using Menu for Density and Fields
